+++ workspaces/persist-layer-diagram/src/Diagram.tsx
@@ -1,0 +1,139 @@
+/**
+ * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ *
+ */
+
+import React, { useEffect, useState } from 'react';
+import { ComponentModel, GetPersistERModelResponse } from '@wso2-enterprise/ballerina-languageclient';
+import { DiagramEngine, DiagramModel } from '@projectstorm/react-diagrams';
+import { CanvasWidget } from '@projectstorm/react-canvas-core';
+import CircularProgress from '@mui/material/CircularProgress';
+import { CMEntity as Entity } from '@wso2-enterprise/ballerina-languageclient';
+import { modelMapper, generateEngine } from './utils';
+import { DiagramControls, HeaderWidget, OverlayLayerModel, PersistDiagramContext, PromptScreen } from './components';
+import { ERRONEOUS_MODEL, NO_ENTITIES_DETECTED, dagreEngine, Colors } from './resources';
+import { Container, DiagramContainer, useStyles } from './utils/CanvasStyles';
+
+import './resources/assets/font/fonts.css';
+
+interface PersistDiagramProps {
+    getPersistModel: () => Promise<GetPersistERModelResponse>;
+    selectedRecordName: string;
+    showProblemPanel: () => void;
+}
+
+export function PersistDiagram(props: PersistDiagramProps) {
+    const { getPersistModel, selectedRecordName, showProblemPanel } = props;
+
+    const [diagramEngine] = useState<DiagramEngine>(generateEngine);
+    const [diagramModel, setDiagramModel] = useState<DiagramModel>(undefined);
+    const [selectedNodeId, setSelectedNodeId] = useState<string>(undefined);
+    const [hasDiagnostics, setHasDiagnostics] = useState<boolean>(false);
+    const [userMessage, setUserMessage] = useState<string>(undefined);
+    const [collapsedMode, setIsCollapsedMode] = useState<boolean>(false);
+
+    const styles = useStyles();
+
+    useEffect(() => {
+        refreshDiagram();
+        const nodeId = selectedRecordName ? `$anon/.:0.0.0:${selectedRecordName}` : '';
+        if (nodeId !== selectedNodeId) {
+            setSelectedNodeId(nodeId);
+        }
+    }, [props]);
+
+    useEffect(() => {
+        if (diagramEngine.getCanvas()) {
+            function handleEscapePress(event: KeyboardEvent) {
+                if (event.key === 'Escape' && selectedNodeId) {
+                    setSelectedNodeId(undefined);
+                }
+            }
+            document.addEventListener('keydown', handleEscapePress);
+        }
+    }, [diagramModel, diagramEngine.getCanvas()]);
+
+    const refreshDiagram = () => {
+        getPersistModel().then(response => {
+            const pkgModel: Map<string, ComponentModel> = new Map(Object.entries(response.persistERModel));
+            const entities: Map<string, Entity> = new Map(Object.entries(pkgModel.get('entities')));
+            setHasDiagnostics(response.diagnostics.length > 0);
+            if (entities.size) {
+                const model = modelMapper(entities);
+                model.addLayer(new OverlayLayerModel());
+                diagramEngine.setModel(model);
+                setDiagramModel(model);
+                autoDistribute(model);
+            } else if (response.diagnostics.length && !diagramModel) {
+                setUserMessage(ERRONEOUS_MODEL);
+            } else if (!response.diagnostics?.length) {
+                setDiagramModel(undefined);
+                setUserMessage(NO_ENTITIES_DETECTED);
+            }
+        });
+    }
+
+    const autoDistribute = (model: DiagramModel) => {
+        setTimeout(() => {
+            dagreEngine.redistribute(diagramEngine.getModel());
+            diagramEngine.zoomToFitNodes({ margin: 10, maxZoom: 1 });
+            diagramEngine.getModel().removeLayer(diagramEngine.getModel().getLayers().find(layer => layer instanceof OverlayLayerModel));
+            diagramEngine.setModel(model);
+        }, 30);
+    };
+
+    const switchCollapseMode = (shouldCollapse: boolean) => {
+        setIsCollapsedMode(shouldCollapse);
+        if (diagramModel) {
+            autoDistribute(diagramModel);
+        }
+    }
+
+    let ctx = {
+        collapsedMode,
+        selectedNodeId,
+        setSelectedNodeId,
+        setHasDiagnostics,
+        hasDiagnostics
+    }
+
+    return (
+        <Container>
+            <PersistDiagramContext {...ctx}>
+                <HeaderWidget collapsedMode={collapsedMode} setIsCollapsedMode={switchCollapseMode} />
+                <DiagramContainer>
+                    {diagramEngine?.getModel() && diagramModel ?
+                        <>
+                            <CanvasWidget engine={diagramEngine} className={styles.canvas} />
+                            <DiagramControls
+                                engine={diagramEngine}
+                                refreshDiagram={refreshDiagram}
+                                showProblemPanel={showProblemPanel}
+                            />
+                        </> :
+                        userMessage ?
+                            <PromptScreen
+                                userMessage={userMessage}
+                                showProblemPanel={hasDiagnostics ? showProblemPanel : undefined}
+                            /> :
+                            <CircularProgress sx={{ color: Colors.PRIMARY }} />
+                    }
+                </DiagramContainer>
+            </PersistDiagramContext>
+        </Container>
+    );
+}