+++ workspaces/statement-editor/src/components/Forms/GraphqlForms/ServiceClassResource.tsx
@@ -1,15 +1,11 @@
-/*
- * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com). All Rights Reserved.
- *
- * This software is the property of WSO2 Inc. and its suppliers, if any.
- * Dissemination of any information or reproduction of any material contained
- * herein is strictly forbidden, unless permitted by WSO2 in accordance with
- * the WSO2 Commercial License available at http://wso2.com/licenses.
- * For specific language governing the permissions and limitations under
- * this license, please see the license as well as any agreement you’ve
- * entered into with WSO2 governing the purchase of this software and any
- * associated services.
- */
+/**
+ * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 LLC. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content."
+ */
 
 // tslint:disable: jsx-no-multiline-js jsx-no-lambda
 import React, { useContext, useEffect, useState } from "react";
@@ -328,7 +324,7 @@ export function ServiceClassResourceForm(props: FunctionProps) {
     };
 
     const onPathFocus = () => {
-        setCurrentComponentName("Path");
+        setCurrentComponentName("Field");
     };
 
     const handleResourceParamChange = async (
@@ -396,13 +392,13 @@ export function ServiceClassResourceForm(props: FunctionProps) {
             <>
                 <div className={connectorClasses.formContentWrapper}>
                     <div className={connectorClasses.formNameWrapper}>
-                        <FieldTitle title="Path" optional={false}/>
+                        <FieldTitle title="Field" optional={false}/>
                         <LiteTextField
                             value={resourceName}
                             onChange={handlePathChange}
                             onFocus={onPathFocus}
-                            isLoading={currentComponentName !== "Path" && isEditInProgress}
-                            diagnostics={(currentComponentName === "Path")
+                            isLoading={currentComponentName !== "Field" && isEditInProgress}
+                            diagnostics={(currentComponentName === "Field")
                             && getPathDiagnostic()}
                         />
                         <Divider className={connectorClasses.sectionSeperatorHR}/>