+++ workspaces/statement-editor/src/components/Forms/Utils/FormUtils.tsx
@@ -1,15 +1,11 @@
-/*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
- *
- * This software is the property of WSO2 Inc. and its suppliers, if any.
- * Dissemination of any information or reproduction of any material contained
- * herein is strictly forbidden, unless permitted by WSO2 in accordance with
- * the WSO2 Commercial License available at http://wso2.com/licenses.
- * For specific language governing the permissions and limitations under
- * this license, please see the license as well as any agreement you’ve
- * entered into with WSO2 governing the purchase of this software and any
- * associated services.
- */
+/**
+ * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ *
+ * This software is the property of WSO2 LLC. and its suppliers, if any.
+ * Dissemination of any information or reproduction of any material contained
+ * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
+ * You may not alter or remove any copyright or other notice from copies of this content."
+ */
 import {
     createFunctionSignature,
     createListenerDeclartion,
@@ -66,15 +62,17 @@ export function getInitialSource(type: string, targetPosition: NodePosition): st
         case "Resource": {
             return getSource(createResource("get", "path", '', "error?", targetPosition));
         }
-        case "GraphqlResource":
+        case "GraphqlResource": {
+            return getSource(createResource("get", "queryName", '', "string", targetPosition));
+        }
         case "ServiceClassResource": {
-            return getSource(createResource("get", "path", '', "string", targetPosition));
+            return getSource(createResource("get", "fieldName", '', "string", targetPosition));
         }
         case "GraphqlMutation": {
             return getSource(createRemoteFunction("mutate", '', "string", targetPosition));
         }
         case "GraphqlSubscription": {
-            return getSource(createResource("subscribe", "path", '', "stream<string>", targetPosition));
+            return getSource(createResource("subscribe", "subscriptionName", '', "stream<string>", targetPosition));
         }
     }
     return;